lfs: avoid deadlock when object scanner returns early
When we encounter an end of file in the object scanner, we can return
false from the Scan method.  In such a case, if there are a sufficient
number of blobs left to be handled, the goroutine that produces them
will fill the channel it uses to send them to us, and since it is not
done, it will close neither the blob channel nor the error channel.

However, when we exit the loop early, we still attempt to wait for the
error channel to close, which means we deadlock.  If we return early,
skip the attempt to read from the error channel and simply assume no
error has occurred.