git: ensure fsck works in a subdirectory
When we look up the object directory for a repository, we do it based on
the common directory, as provided by git rev-parse
--git-common-directory. In recent versions of Git, this is relative to
the current directory. However, before Git 2.13.0, this output was
broken, and in subdirectories provided a nonsensical value.

Since we computed the value based on the current directory, git lfs fsck
failed in a subdirectory with a missing object error when using a Git
before 2.13.0. Ensure that the code works correctly in both cases by
resolving the path against the repository root on older versions of Git.

Add a test to ensure this works on whatever version of Git people happen
to be using.