Turn on long path support for Windows 10
On Windows, by default, a path is limited to 260 characters. It is
possible to use UNC-style paths, but these can only be absolute.
However, Windows 10 has introduced long path support if the application
supports it by providing a manifest and the user has enabled it. Specify
a manifest on Windows so that we can support long paths.

Note that the Go standard library already avoids using fixed-size
buffers for paths, so it should handle arbitrarily long path names
without a problem.