t: unset test harness variables
We set some variables for the test harness when running in CI on
Windows.  However, if these variables leak into the environment of the
Git LFS process, they'll cause the test to fail, since they'll be
printed in the env output.  Make sure that we unset them in the tests
that care about this to ensure that they don't cause failures.