t: improve locking tests
Currently, our locking tests test how we behave when a file is listed in
.gitignore and assume that we're testing behavior with ignored files.
However, because the file is already checked in, we're not testing
actual ignored files and locking.

Instead, let's test how locking behavior works for files that are
actually ignored and not checked in at all.  This tests the behavior we
care about, both with and without the relevant option.