WIP: tools: use git ls-files to speed up tree walks
Currently, our tree walks are quite slow and our filepath filtering
code, which is used by the tree walks, is not 100% accurate with respect
to .gitignore files.  To address both of these issues, use git ls-files
to find the files to iterate over.  This means that we get to take
advantage of the speed of the index and additionally get correct
.gitignore handling.