CHANGELOG: replace "master" with "main"
The term "master" is problematic due to its association with
master/slave terminology.  Since we're renaming the default branch to
"main", let's update the changelog.  Note that in addition to the link,
which must be updated for correctness, we update historical output since
it has the same problematic terminology and doing so is easy.