script/cibuild: avoid running go generate when formatting
When we're running "make fmt" at the end of the CI build to check
formatting, we don't want to run "go generate", since we're not building
anything.  However, placing the additional environment variables in the
middle of the command invokes the default target and does just that.
Let's avoid this by placing the environment variable at the end after
the fmt target.

Note that this doesn't negatively affect other places we do this in the
cibuild script, since we're either already building or are intentionally
building again to test a particular corner case.