script/cibuild: avoid running lint rule when checking formatting
Running the lint rules causes dependencies to be downloaded and built,
which causes problems in making the tree dirty on Windows.  We just
want to run goimports with a suitable set of arguments, so let's pretend
the lint target was just built with -W, which prevents it from being
run.